Consolidate AI tool calling subsystem into AITools.swift
